picasa/requirements.txt
PyQt6>=6.6.0
Pillow>=10.0.0
piexif>=1.1.3
rawpy>=0.18.0
imageio>=2.33.0
imageio-ffmpeg>=0.4.9
folium>=0.15.0
PyQt6-WebEngine>=6.6.0
requests>=2.28.0
